+++ news.md
@@ -3,4 +3,4 @@
 - *[IMPROVED]:* handling of "better animal welfare" icon instead of outputting the confusing beef icon
 - *[IMPROVED]:* under the hood improvements
 
-- *[FIXED]:* English menu via `/mensaEN`
+- *[FIXED]:* English menu via /mensaEN